requirements.txt
python-dotenv>=1.0.0
apify-client>=1.6.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
twilio>=8.0.0
requests>=2.28.0
